refactoring: extracting utils funcs to extra file, type fixing and linting
